+++ java/ql/test/kotlin/library-tests/maps-iterator-overloads/test.expected
@@ -1,0 +1,2 @@
+| test.kt:1:100:1:109 | mutableIterator(...) | mutableIterator |
+| test.kt:3:73:3:82 | iterator(...) | iterator |
