+++ java/ql/test/kotlin/library-tests/maps-iterator-overloads/test.kt
@@ -1,0 +1,3 @@
+fun getIter(m: MutableMap<String, Int>): MutableIterator<MutableMap.MutableEntry<String, Int>> = m.iterator()
+
+fun getIter2(m: Map<String, Int>): Iterator<Map.Entry<String, Int>> = m.iterator()
