+++ java/ql/test/kotlin/library-tests/maps-iterator-overloads/test.ql
@@ -1,0 +1,4 @@
+import java
+
+from MethodAccess ma
+select ma, ma.getCallee().toString()
